+++ qiskit_nature/operators/second_quantization/spin_op.py
@@ -388,6 +388,17 @@ class SpinOp(SecondQuantizedOp):
         return self.simplify(atol=atol)
 
     def simplify(self, atol: Optional[float] = None) -> "SpinOp":
+        """Simplify the operator.
+
+        Merges terms with same labels and eliminates terms with coefficients close to 0.
+        Returns a new operator (the original operator is not modified).
+
+        Args:
+            atol: Absolute tolerance for checking if coefficients are zero (Default: 1e-8).
+
+        Returns:
+            The simplified operator.
+        """
         if atol is None:
             atol = self.atol
